Factor out `paginate`
